+++ src/Main.java
@@ -21,102 +21,10 @@ class App {
         System.out.println(firstBook.equals(secondBook));
         System.out.println(secondBook.equals(firstBook));
 
+        System.out.println(firstAuthor.equals(secondAuthor));
+        System.out.println(secondAuthor.equals(firstAuthor));
+
     }
 }
 
 
-class Author {
-
-    String firstName;
-    String secondName;
-
-    public Author() {
-        firstName = "Undefined";
-        secondName = "Undefined";
-    }
-
-    public Author(String firstName, String secondName) {
-        this.firstName = firstName;
-        this.secondName = secondName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public void setSecondName(String secondName) {
-        this.secondName = secondName;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public String getSecondName() {
-        return secondName;
-    }
-
-    public String toString() {
-        return firstName + " " + secondName;
-    }
-
-    public boolean equals(Object obj){
-        return obj instanceof Author;
-    }
-
-    public int hashCode() {
-        return 10*(this.toString()).hashCode();
-    }
-
-}
-
-class Book {
-
-    String name;
-    Author author;
-    int year;
-
-    public Book(String name, Author author, int year) {
-        this.name = name;
-        this.author = author;
-        this.year = year;
-    }
-
-    public void setYear(int year) {
-        this.year = year;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public void setAuthor(Author author) {
-        this.author = author;
-    }
-
-    public int getYear() {
-        return year;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getAuthor() {
-        return author.getFirstName() + " " + author.getSecondName();
-    }
-
-    public String toString() {
-        return author.toString() + " " + year + " \"" + name + "\"";
-    }
-
-    public boolean equals(Object obj){
-        return obj instanceof Book;
-    }
-
-    public int hashCode() {
-        return 10 * author.hashCode() + 1;
-    }
-
-
-}
